requirements.txt
python-telegram-bot[all]>=21.0,<22.0
openai>=1.0.0,<2.0.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
aiohttp>=3.9,<4.0
stripe>=7.0,<9.0
redis[hiredis]>=5.0,<6.0
